requirements.txt
anthropic>=0.40.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
